// src/app/modules/admin/modules/admin-items/models/edit-item.ts
export interface EditItem {
    id: number;
    active: boolean;
    name: string;
    description: string;
    color: string;
    brand: string;
    price: number;
    discount: number;
    kind: number;
    subkind: number;
    status: number;
    size: string;
    amount: number;
    sex: number;
    previewImagePath: string;
    imagePath1: string;
    imagePath2: string;
    imagePath3: string;
}